+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * update dependencies
 * add token_customset to return parameter of privilegekeylist for server version 3.11
 * add -banners flag to channellist method and added getters to the TeamSpeakChannel class
+* ResponseError has now the property `source` which holds the Command instance which generated the error
 
 ### 2.3.1 - 13.12.2019
 * added more exports for the index file